Enable dotSize in the MDS plot to be adjustable
In the original version, dotSize in the MDS plot could not be adjusted by users.
Reason identified:
The dotSize parameter is stored in userOption instead of localOptions, but the old script mistakenly tried to retrieve the dotSize information from the localOption. I edited the code so that dotSize is now rerived from the correct variable.