Print out MDS comment if it failed at the initial trial
By design, when MDS failed at the initial trial, it might be fixed (in the catch blocks) by changing criterion from user-specified criterion to 'stress', or adding 0.2 to distances to avoid colocalization. However, this information is never shown or accessible by users. That is, MDS plot is displayed as normal and users could easily ignore the fact that the MDS criterion might have been changed. Therefore, I edit the codes so that whenever the MDS fails at the initial trial (and is fixed afterwards by the catch blocks), the description of the changes made is visible to the users.