Pre-weighten scaling by Bcov, not resMS
This is an important change that effects the scaling of the crossnobis
distance (LDC). In previous version we had simply scaled the variance
matrix to reflect the variance of the time series. This can lead to
instabilities of the distance across multiple sessions (but not for a
single session), especially when using RWLS to estimate the
variance-covariance matrix. The current fix should remove these
instabilities. I makes the pre whitened beta weights to have variance I
in the spatial direction, not variance mean(diag(Bcov)). For most
cases, you will only see that the overall size of the distances is
substantially larger than before - and not notice any difference in any
other results.